// src/types/ImageSettings.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Corner } from "./Corner";

export type ImageSettings = { inputDirectory: string, outputDirectory: string, clearFilesInputDirectory: boolean, searchChildFolders: boolean, clearFilesOutputDirectory: boolean, keepChildFoldersStructureInOutputDirectory: boolean, overwriteExistingFilesOutputDirectory: boolean, minPixelCount: number, addLogo: boolean, logoPath: string | null, logoScale: number, logoXOffsetScale: number, logoYOffsetScale: number, logoCorner: Corner, shouldConvertFormat: boolean, format: string, };